requirements.txt
requests
pandas
pymongo
python-dotenv
beautifulsoup4
lxml
selenium
webdriver-manager
sqlalchemy
psycopg2-binary